+++ packages/admin/client/views/pages/admin.js
@@ -1,10 +1,51 @@
 
 Template.pageAdmin_admin_raw.helpers({
 	rawdata: function() {
+		var data = {};
+
 		if(this.type==='user')
-			return Users.findOne(this._id);
+			data = Users.findOne(this._id);
 		else if(this.type==='place')
-			return Places.findOne(this._id);
+			data = Places.findOne(this._id);
+
+		return JSON.stringify(data,null,4);
+	}
+});
+
+Template.pageAdmin_admin_raw.events({
+	'shown.bs.collapse ': function(e,tmpl) {
+		tmpl.$('.btn-edit').removeClass('hidden');
+	},
+	'hidden.bs.collapse ': function(e,tmpl) {
+		tmpl.$('.btn-edit').addClass('hidden disabled');
+	},
+	'keyup .text-data': function(e,tmpl) {
+		tmpl.$('.btn-edit').removeClass('disabled');
+	},
+	'click .btn-edit': function(e,tmpl) {
+		var ico$ = $(e.target).find('.icon'),
+			text = tmpl.$('.text-data').val();
+
+		try {
+
+			data = JSON.parse(text);
+		}
+		catch(e) {
+			console.log(data)
+			K.Alert.warning(i18n('error_admin_rawedit'));
+			return false
+		}
+
+		if(tmpl.data.type==='place') {
+			ico$.addClass('icon-loader');
+			K.Admin.call('updatePlace', tmpl.data.id, data, function(res) {
+				ico$.removeClass('icon-loader');
+			});
+		}
+		/*TODO else if(tmpl.data.type==='user')
+			K.Admin.updateUser(tmpl.data.id, data, function(res) {
+				ico$.removeClass('icon-loader');
+			});*/
 	}
 });